Add a token generator script
