+++ src/google/google-doc.ts
@@ -1,6 +1,6 @@
 import * as fs from 'fs';
 import { Auth, docs_v1, google } from 'googleapis';
-import { getNewToken } from './create-token';
+import { getNewToken } from './token-generator';
 
 // The file token.json stores the user's access and refresh tokens, and is
 // created automatically when the authorization flow completes for the first
